# docs/swagger.yaml
definitions:
...
  model.BaseResponse:
    properties:
      errCode:
        description: 错误码
        type: string
      errMessage:
        description: 错误信息
        type: string
      success:
        description: 请求是否成功
        type: boolean
    type: object
...
  model.File:
    properties:
      created_at:
        type: string
      file_data:
        items:
          type: integer
        type: array
      file_id:
        type: integer
      file_name:
        type: string
      file_size:
        type: integer
      file_type:
        type: string
      location:
        type: string
      related_id:
        type: integer
      updated_at:
        type: string
    type: object
  model.FileReqCreate:
    properties:
      file_data:
        items:
          type: integer
        type: array
      file_name:
        type: string
      file_size:
        type: integer
      file_type:
        type: string
      location:
        type: string
      related_id:
        type: integer
    required:
    - file_data
    - file_name
    - file_type
    - location
    - related_id
    type: object
  model.FileReqEdit:
    properties:
      file_data:
        items:
          type: integer
        type: array
      file_id:
        type: integer
      file_name:
        type: string
      file_size:
        type: integer
      file_type:
        type: string
      location:
        type: string
      related_id:
        type: integer
    required:
    - file_id
    type: object
  model.FileReqList:
    properties:
      keyword:
        type: string
      page:
        type: integer
      page_size:
        type: integer
    required:
    - page
    - page_size
    type: object
...
  model.ListResponse-model_File:
    properties:
      errCode:
        description: 错误码
        type: string
      errMessage:
        description: 错误信息
        type: string
      list:
        description: 列表
        items:
          $ref: '#/definitions/model.File'
        type: array
      success:
        description: 请求是否成功
        type: boolean
      total:
        description: 总条数
        type: integer
    type: object
  model.ListResponse-model_Menu:
    properties:
      errCode:
        description: 错误码
        type: string
      errMessage:
        description: 错误信息
        type: string
      list:
        description: 列表
        items:
          $ref: '#/definitions/model.Menu'
        type: array
      success:
        description: 请求是否成功
        type: boolean
      total:
        description: 总条数
        type: integer
    type: object
  model.ListResponse-model_Notice:
    properties:
      errCode:
        description: 错误码
        type: string
      errMessage:
        description: 错误信息
        type: string
      list:
        description: 列表
        items:
          $ref: '#/definitions/model.Notice'
        type: array
      success:
        description: 请求是否成功
        type: boolean
      total:
        description: 总条数
        type: integer
    type: object
...
  model.Menu:
    properties:
      created_at:
        type: string
      display_order:
        type: integer
      is_active:
        type: boolean
      menu_code:
        type: string
      menu_id:
        type: integer
      menu_name:
        type: string
      updated_at:
        type: string
    type: object
  model.MenuReqCreate:
    properties:
      display_order:
        type: integer
      is_active:
        type: boolean
      menu_code:
        type: string
      menu_name:
        type: string
    required:
    - is_active
    - menu_code
    - menu_name
    type: object
  model.MenuReqEdit:
    properties:
      display_order:
        type: integer
      is_active:
        type: boolean
      menu_code:
        type: string
      menu_id:
        type: integer
      menu_name:
        type: string
    required:
    - menu_id
    type: object
  model.MenuReqList:
    properties:
      keyword:
        type: string
      page:
        type: integer
      page_size:
        type: integer
    required:
    - page
    - page_size
    type: object
  model.Notice:
    properties:
      content:
        type: string
      created_at:
        type: string
      is_active:
        type: boolean
      is_read:
        type: boolean
      notice_id:
        type: integer
      notice_type:
        type: boolean
      published_at:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  model.NoticeReqCreate:
    properties:
      content:
        type: string
      is_active:
        type: boolean
      is_read:
        type: boolean
      notice_type:
        type: boolean
      published_at:
        type: string
      title:
        type: string
      user_id:
        type: integer
    required:
    - content
    - is_active
    - notice_type
    - published_at
    - title
    type: object
  model.NoticeReqEdit:
    properties:
      content:
        type: string
      is_active:
        type: boolean
      is_read:
        type: boolean
      notice_id:
        type: integer
      notice_type:
        type: boolean
      published_at:
        type: string
      title:
        type: string
      user_id:
        type: integer
    required:
    - notice_id
    type: object
  model.NoticeReqList:
    properties:
      keyword:
        type: string
      page:
        type: integer
      page_size:
        type: integer
    required:
    - page
    - page_size
    type: object
...
  model.Response-model_File:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/model.File'
        description: 数据
      errCode:
        description: 错误码
        type: string
      errMessage:
        description: 错误信息
        type: string
      success:
        description: 请求是否成功
        type: boolean
    type: object
  model.Response-model_Menu:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/model.Menu'
        description: 数据
      errCode:
        description: 错误码
        type: string
      errMessage:
        description: 错误信息
        type: string
      success:
        description: 请求是否成功
        type: boolean
    type: object
  model.Response-model_Notice:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/model.Notice'
        description: 数据
      errCode:
        description: 错误码
        type: string
      errMessage:
        description: 错误信息
        type: string
      success:
        description: 请求是否成功
        type: boolean
    type: object
...
info:
  contact: {}
paths:
...
  /api/files:
    post:
      consumes:
      - application/json
      description: 新建一个文件
      parameters:
      - description: 文件信息
        in: body
        name: file
        required: true
        schema:
          $ref: '#/definitions/model.FileReqCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response-model_File'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.BaseResponse'
      summary: 新建文件
      tags:
      - Files
    put:
      consumes:
      - application/json
      description: 更新文件信息
      parameters:
      - description: 文件信息
        in: body
        name: file
        required: true
        schema:
          $ref: '#/definitions/model.FileReqEdit'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.BaseResponse'
      summary: 更新文件
      tags:
      - Files
  /api/files/{file_id}:
    delete:
      consumes:
      - application/json
      description: 删除一个文件
      parameters:
      - description: 文件ID
        in: path
        name: file_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.BaseResponse'
      summary: 删除文件
      tags:
      - Files
    get:
      consumes:
      - application/json
      description: 获取单个文件信息
      parameters:
      - description: 文件ID
        in: path
        name: file_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response-model_File'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.BaseResponse'
      summary: 获取文件
      tags:
      - Files
  /api/files/list:
    post:
      consumes:
      - application/json
      description: 获取文件分页列表
      parameters:
      - description: 分页与搜索
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/model.FileReqList'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ListResponse-model_File'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BaseResponse'
      summary: 获取文件列表
      tags:
      - Files
  /api/menus:
    post:
      consumes:
      - application/json
      description: 新建一个菜单
      parameters:
      - description: 菜单信息
        in: body
        name: menu
        required: true
        schema:
          $ref: '#/definitions/model.MenuReqCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response-model_Menu'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.BaseResponse'
      summary: 新建菜单
      tags:
      - Menus
    put:
      consumes:
      - application/json
      description: 更新菜单信息
      parameters:
      - description: 菜单信息
        in: body
        name: menu
        required: true
        schema:
          $ref: '#/definitions/model.MenuReqEdit'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.BaseResponse'
      summary: 更新菜单
      tags:
      - Menus
  /api/menus/{menu_id}:
    delete:
      consumes:
      - application/json
      description: 删除一个菜单
      parameters:
      - description: 菜单ID
        in: path
        name: menu_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.BaseResponse'
      summary: 删除菜单
      tags:
      - Menus
    get:
      consumes:
      - application/json
      description: 获取单个菜单信息
      parameters:
      - description: 菜单ID
        in: path
        name: menu_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response-model_Menu'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.BaseResponse'
      summary: 获取菜单
      tags:
      - Menus
  /api/menus/list:
    post:
      consumes:
      - application/json
      description: 获取菜单分页列表
      parameters:
      - description: 分页与搜索
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/model.MenuReqList'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ListResponse-model_Menu'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BaseResponse'
      summary: 获取菜单列表
      tags:
      - Menus
  /api/notices:
    post:
      consumes:
      - application/json
      description: 新建一条通知
      parameters:
      - description: 通知信息
        in: body
        name: notice
        required: true
        schema:
          $ref: '#/definitions/model.NoticeReqCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response-model_Notice'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.BaseResponse'
      summary: 新建通知
      tags:
      - Notices
    put:
      consumes:
      - application/json
      description: 更新通知内容
      parameters:
      - description: 通知信息
        in: body
        name: notice
        required: true
        schema:
          $ref: '#/definitions/model.NoticeReqEdit'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.BaseResponse'
      summary: 更新通知
      tags:
      - Notices
  /api/notices/{notice_id}:
    delete:
      consumes:
      - application/json
      description: 删除一条通知
      parameters:
      - description: 通知ID
        in: path
        name: notice_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.BaseResponse'
      summary: 删除通知
      tags:
      - Notices
    get:
      consumes:
      - application/json
      description: 获取单个通知信息
      parameters:
      - description: 通知ID
        in: path
        name: notice_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response-model_Notice'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.BaseResponse'
      summary: 获取通知
      tags:
      - Notices
  /api/notices/list:
    post:
      consumes:
      - application/json
      description: 获取通知分页列表
      parameters:
      - description: 分页与搜索
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/model.NoticeReqList'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ListResponse-model_Notice'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BaseResponse'
      summary: 获取通知列表
      tags:
      - Notices
...
swagger: "2.0"